Add command `!log` to pretty-print log to a file.
Many improvements in general.

Tried to improve keeping a record of tokens by the api response but they will vary, hopefully, just a little from the actual token count.

If history file (the tsv file) has got a zero or negative token count, we shall not skip it anymore, as long as there is an actual string saved there. In those cases, we use `__tiktokenf` fun to get an approximate number of tokens for that history entry.