Improve Compat with Zsh.
We recenetly changed `emulate -R zsh` to `emulate zsh` in order to preserve debbuging options such as when the script is run with `zsh -xv chatgpt.sh`.

Well, as we use color codes with `vared`, setting the option PROMPT_PERCENT is required when Zsh is called as the Ksh interpreter.