Fix context generation, max_tokens implementation
In this realease, we (try to) fix a bad implementation of the maximum tokens settings. This settings refers to the **maximum response tokens**, not the **maximum model tokens**.

**Max model tokens** is now set automatically (we always want to use all the tokens available in a model). Setting `-1024` will set **max response tokens**. To set max **response** and **model** tokens, set an option like `-4096-1024`, `-4096,1024` or even `-M 4096-1024` to set model max to 4096 and response max to 1024.

*OBS*: other chatgpt wrappers in GitHub implement *max tokens* as the maximum number of input+response tokens!

Also, context generation for chat modes were not working properly. For ex, the number of tokens of the INSTRUCTION prompt was being added twice to the token stack, it is now fixed.

Added +64 tokens to the context generator to be sure we will never (or almost never) run out of tokens for the reply.

Ksh2020 was tested and it lacks some Ksh93u+ features, such as a good `read command` with history support, for example. Dont use ~~Ksh2020~~ if you can! Prefer **Ksh93u+m**, which is being actively developed!

Updated help page to reflect how the script works. Version bumped to v0.10.0.