Improve defaults chat interface, less verbose.
Fix a bug with KSH `read` when printing PS1 (such as `Q: ` or a restart sequence) to stderr and then pressing up arrow key to fetch a long history item (multiline) that made the textt become mangled.

In BASH, that seems to work all right but as ZSH uses `vared`, it is best if we unify the different interfaces as closest as possible.

Add some new lines for sanity.

The new defaults prompter is minimal and is printed to stderr. When the user type his prompt, it will erase the PS1 (prompt string such as `Q: `).