Improve record of token usage in history file when `Best_of` is set.
Best_of option will consume a lot more tokens when it is set to a value larger than *1*. So we cannot simply do arithmetics with the results from previous prompts to get token values for current items to be record in history because the error is too big (if we simply dive used tokens by Best_of).

In that case, we defaults to getting a minimum token count of the string to be record with our poor mans tiktoken fun. Token count **will compensate** itself with new rounds of server responses in the history file!

***Just to be clear***, token count for single items in history file may be only approximate but total token count should be much accurate.