Improve option -f (ignore user config file and environment).
`Option -f` is useful for debbuging this script (as are options -VVj).

`Option -f` needed updating since many options changed. All options that can be set by the config file and environment should be reset before resourcing the script. Plus, $OPTIND was not being reset (`OPTIND=1`) before resourcing the script, so that is a wonder it was working more or less the way it is supposed to...