Add input multiline support with `read` (chat mode).
Update the new function a little. Iniput example will add new lines between WORD1 and WORD2:

WORD1\
\
\
\
WORD2

Note that if no new lines are needed, it is  not necessary to type in backslash `\\` to keep in multiline mode as it gets activated by the first backslash and stays on until pressing ENTER on an empty line!