Awesome-chatgpt-prompt integration. Fix regresison bug with Zsh option handling.
Set or search for a prompt from awesome-chatgpt-prompt project from GitHub. It downlaods the csv file with awesome prompts and sets it automatically. user may set from the list.

OBS: Zsh does not print all the prompts from awesome-chatgpt-prompt to the terminal when the list exceeds screen size.

Fix serious regressions bug (unsetting OPTARG is a no-no). So that option handling was not working correctly in Zsh and possibly in Bash, as well.

Change `test -e` to `test -f` to avoid a directory `/` returning true.