Avoid deleting older entries with `!regen` when response errors out.
So, if curl cannot get a response from the server, flag as `$BAD_RESPONSE`, so on trying to regenerate a response with `!regen`, the script will not `sed $d` twice the history file, or if "BREAK SESSION" flag is set on the history file.

Some small improvements with chat commands in general.