Important fixes for breaking from loop and recording to hist with no `-ccCC`.
In recent update, we added `$BAD_RESPONSE` variable to catch when an error is thrown from the server, useful to avoid removing history lines with subsequent `!regenerate` command until a valid response arrives. It was also causing single-run mode to not break from the loop!

Also, without `-ccCC` options, the script would not record a restart and start sequence to history file if none set by the user, which would mess up when reading from history (the script would not know if that was input or response)!