Version 0.12. Many important fixes and much improved code.
Killed old code which attempted to change interlocutor types on the run, but it became evident it is almost useless and just added a lot of complexity to the code.. There are so many good ideas to implement!

Fixed recording to the history file. We should always have the type (or role) of the message in the history file. So we just use our `\\nQ:\ ` and `\\nA:` types and if user wants setting Restart and Start sequences, then we reload the history file and add those instead of our fixed types.

Change option -CC to resemble PLayGround more in pure text completions.. Even a space here can make a difference!

Working with the models in pure text completions -CC, it should be required setting Restart Start sequences manually, so we now respect that to have a more close experience to web PlayGround... A space is added to the input prompt to make sure user notices a space is important if no restart and start seqs are set in pure text cmpls!

Option -HH pretty prints last chat history as usual, but if combined with -C, we don't print the chat mode defaults Restart Start sequences (only if the user set those two seqs manually), so it is more like natural text.

We are defaulting to be a little more verbose with the chat mode options.

Vars `Q_TYPE` and `A_TYPE` should not be easily available to user configs because we mostly use them intenally, and to record entries with a role/type to history file.

Renamed some variables.

Changed defaults shell to `bash`. Sorry guys, `ksh` interative features still have some bugs and `bash` is no match in those respects..
A lot of code had to be rewritten to use extended globbing with moderation as extended globbing seems to make `bash` run soo slow compared to `ksh` and `zsh`. I am sure `bash` extended globbing development has been left neglected as most people use regex instead.

Removed compatibility with Ksh and Zsh.

Moved detailed explanation to the manual page and rewrote the help page.