Fix slash command to regenerate last answer.
The case statement in the multiline read prompter would try and check input for a command, however, not only it checked the command as it executed it, too! So, just removing it.

Also, adding new code to try and tokenize input. It uses OpenAI tiktoken libs... For now, we are still using our poor man __tiktokenf function in the chatgpt script...