Improve `options -HH and -L`.
Uses sed+cat to format new lines in a pretty way for user log instead of var substitution..
The thing is, it was darn fast with Ksh and Zsh, but Bash is soo slow, so we cannot rely on it for some more demanding processing... In retrospective, Zsh could have been chosen as the defaults interpreter... So, we may need to use Zsh in the future if need be...

The only problem with Zsh is history is not available, but there are things we can do to overcome that.

Anyways, I digress... We will be using Bash for a while.