Improve history management for Bash and Zsh.
With option -Z, try to restart script with `zsh -if $0` with interactive mode, without reading cinfig files.
This way, we can use native Zsh history system, which is much faster than trying to load it manyally.

However, the code to load history manually still persists, in case zsh is not started interactivelly.

Check when the history file has changed. If changed, prompt user for a new session name.