Option -y, count tokens accurately when generating history context.
The script uses a hack to try and preview token count in some instances, however token count changes a lot from model encoding to encoding.

The history file actually records the number of tokens for each prompt accurately using an ingenious way, if I may say so myself, to get accurate token count using the difference with previous token count returned.. Only the first prompts in the history file may have disparating token count: the first prompt, or the instruction prompt, is recorded as a preview token count as we dont have access to a previous answer. However, this preview is compensated by further responses as we can readily use simple arithmetics to get the exact number of tokens after we receive a response.

Now, when trying to continue a conversation from history, it will add up all the recorded tokens, if using the history with a different model or encoding, the history context length may be wrong.

In such cases, option -y will make use of python tiktoken to recount the number of tokens for each history entry and thus keep a  context within token limits.

The caveat: it is slow as we need to re-run the little python script for each history entry read.

--

Also, add chat command `!fork` synonymous to `!copy`.
improve help and man pages with long options lumped together with the short options.