Push INSTRUCTION to script shell history. Improve handling of new lines in text cmpls.
That is useful to have the instruction pushed to the script shell history, specially when using awesome-chatgpt-prompts and editing it before the first request!

Also, improve handling of new lines with the plain/pure text completions mode, that is beware of adding new lines to input automatically as new lines are tokens which should be heeded in text completions carefully!

Check response to see if a space is warranted or not in the new prompt (is continuation of a line or was there a new line in response?).

Setting the context from the history file removed trailinig new lines of the initial INSTRUCTION prompt when that was the only prompt in the context retrieved from the history session. Following turns would prepare the instruction and following prompts all righty, however removal of trailing new lines also ocurred in some other instances, for example when the response contained multiple new lines. That is desirable in the chat mode, however that is not desired a behaviour when using plain text completions in multi-turn (`option -d`) because the new lines may be intended.

Note that removing trailing and leading new lines and white spaces is still done and *is* desirable in chat mode (as opposed to plain text completions). In the future, it may be implemented that all but the first 2 trailing new lines should be deleted in plain text completions (some newlines may be meaningful but too many new lines are usually not desirable even in plain text completions and may result in lower quality prompts).

Setting `$OPTS` (image size, Large, Medium or Small) from the script configuration file would prevent setting image size from the first positional paramater as described in the man page. That is now fixed.

Also, restrict reading the second pos arg as image size in some cases (do note image size should always be the first pos arg with `option -i`)!