Fix setting output image size with first pos arg with `option -i` (Small, Medium, or Large).
The last commit introduced an error when setting `$OPTS` in the script config file as `Small`, `Medium` and `Large`, it was not being converted to `256x256`, `512x512`, and `1024x1024` anymore. That is now fixed.

Check input image colour space, it must be RGB(A), sRGB is not allowed so we convert to PNG32 standards.

Improve verbose warning with `option -i` about entering Image Generations, or Image Edits.